[lib/framework] refactor: move finding gotest subtests to operations mod
