+++ lib/framework/golang/treesitter/operations/mod.rs
@@ -1,2 +1,3 @@
+pub(crate) mod extract_gotest_subtests;
 pub(crate) mod get_build_tags;
 pub(crate) mod parse_tree;
